requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
black==23.7.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
exceptiongroup==1.1.2
fastapi==0.100.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
netifaces==0.10.6
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
protobuf==4.21.12
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
PyMySQL==1.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
sniffio==1.3.0
socketio==0.2.1
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
websockets==11.0.3
Werkzeug==2.3.6
yarl==1.9.2
